% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoRF.R
\name{AUC.randomForest.Pred}
\alias{AUC.randomForest.Pred}
\title{Calculate the auc of RF predictions}
\usage{
AUC.randomForest.Pred(PredResp)
}
\arguments{
\item{PredResp}{A vector of predictions (fraction of votes) and the predicted response vector binded together, e.g. cbind(Prediction,Response).}
}
\value{
Returns the auc.
}
\description{
Calculate the auc of a RF using a set of RF prediction. Primarly useful for cross-validated predictions or predictions made on independent data.
}

