% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoRF.R
\name{AUC.randomForest}
\alias{AUC.randomForest}
\title{Calculate the oob-auc of a rfsrc object}
\usage{
AUC.randomForest(rf)
}
\arguments{
\item{rf}{A rfsrc object.}
}
\value{
Returns the oob-auc.
}
\description{
Calculate the oob-auc of a rfsrc object (from randomForestSRC) using rfsrc$predicticed.oob and the binary response variable.
}

