% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoRF.R
\name{cvCoRF}
\alias{cvCoRF}
\title{n-fold cross-validation of CoRF}
\usage{
cvCoRF(CoRFObject, nfold = 10, SeedCVO = 123, GammaSeq = NULL,
  ntree = NULL, SelectForest = TRUE)
}
\arguments{
\item{CoRFObject}{A CoRF object.}

\item{nfold}{The cross-validation fold.}

\item{SeedCVO}{The seed used to draw the cross-validation folds.}

\item{GammaSeq}{Set gamma sequence to be used in the cross-validation. By default uses the same sequence as used to fit the CoRF object.}

\item{ntree}{Set the number of trees used per RF in the cross-validation. By default uses the same number of trees as used to fit the CoRF object.}
}
\value{
Returns An object with the result of the cross-validation, containng the following components.
\item{foldid}{The cross-validation folds used}
\item{OptGammaPerFold}{When more than one gamma was used, this returns per fold the gamma with lowest brier/auc.}
\item{Predictions}{The cross-validated predictions, for the base RF, the CoRF with gamma=1, and optionally the predictions for the CoRF selected by brier or auc.}
\item{cvResult}{The result of the cross-validation, e.g. cross-validated auc, brier, and accuracy.}
}
\description{
Conducts a n-fold cross-validation of a CoRF object to obtain an unbiased estimate of the predictive error. Primarly for when gamma is tuned.
}

